Update README.md
As mentioned before, operational efficiency is a key metric in estimating performance of business units [5]. However, due to challenges such as subjectivity and lack of standardization, it becomes difficult to perform any benchmarking for business units and thereby figure out slacks, if any, in the performance of the units.
Ordinarily for a retail chain, efficiency may be understood as a ratio of the outputs of a business process to the inputs provided [8]. Inputs may be defined as resources such as capital, labor, infrastructure (floor space), target population with identified catchment area, number of product categories and SKUs, whereas outputs may be defined as revenue (sales value), acquired customers (unique), number of transactions, volume of items sold etc.
However, when multiple inputs and outputs are involved (as is the case in any practical business scenario), estimating efficiency becomes a complicated process. This in turn prevents management from identifying underperforming business units and take remedial measures or redistribute resources [9].
The specific case here refers to 30 stores in the state of Ohio in North America for a multi-brand retail chain for the purpose of the paper. The original project included other states and hence more stores. Regional census data from public data sources along with operations' metrics from the stores have been factored in as well. Certain inputs and output attributes have been prioritized with the help of correlation along with analytical hierarchical process (AHP) method. From this, efficiencies of stores have been identified and analyzed, keeping in mind current business constraints and regional target groups. Multiple algorithms have been deployed for the above analysis, including but not limited to Voronoi Tessellation, Data Envelopment Analysis among other methods.